Huge folder refactor and files renaming
