+++ src/lib/assetManager.mjs
@@ -1,5 +1,5 @@
-import './vendor/GLTFLoader.js';
-import './vendor/OBJLoader.js';
+import '../vendor/GLTFLoader.js';
+import '../vendor/OBJLoader.js';
 
 const BASIS_LIB_PATH = '../src/vendor/';
