+++ src/lib/text.mjs
