+++ src/stations/Paintings.mjs
