+++ src/worlds/PhotogrammetryObject.mjs
